introduces new component "readonly_objects"
If enabled and available, the method
'readonly_object?' will be used to check
wether or not all field of an object
should be readonly.

Useful if forms are used not only to edit an object,
but to just show content of an object. Or if the
object can only be edited in certain states.